requirements.txt
alembic==1.12.0
blinker==1.6.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing_extensions==4.8.0
Werkzeug==2.3.7
WTForms==3.0.1
